feat: change profile picture to drive off filename vs personId
BREAKING CHANGE:  profile-picture personId property has been removed. Use fileName to supply profile picture file name off the CDN.  link-to-profile was replaced by profile-picture-link-person-id where a person id must be supplied to construct a valid like to directory.  All cache busting was removed because it is no longer needed with this approach.  A profilePictureFileName was added to user manager to supply picture file name of logged in user.